Add types to its own file
