+++ src/hooks/useKanaQuizList.tsx
@@ -1,22 +1,7 @@
 import { useState } from 'react';
 import { KanaType } from '../types/kana';
 import { sample, sampleSize } from 'lodash';
-
-// Becuase of Lodash. Can I overwrite the Lodash Types to get rid of the undefined?
-type Answer = KanaType | undefined;
-type Question = KanaType | undefined;
-type CurrentQuestion = { question: Question; answers: Answer[] };
-type QuizStats = { right: number; wrong: number; tries: number };
-
-type HookReturn = {
-  quizKanaLength: number;
-  isDisabled: boolean;
-  currentQuestion: CurrentQuestion;
-  quizStats: QuizStats;
-  feedback: string;
-  getNewQuestion: () => void;
-  checkAnswer: (answer: Answer) => void;
-};
+import { Answer, CurrentQuestion, HookReturn } from '../types/kanaQuiz';
 
 export default function useKanaQuizList(kana: KanaType[]): HookReturn {
   const [quizKana, setQuizKana] = useState(kana);
